pyproject.toml
[project]
name = "redegades-forecast-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.10,<4",
    "fastapi>=0.111,<1",
    "mcp>=1,<2",
    "numpy>=1.26,<3",
    "pydantic>=2.7,<3",
    "python-dotenv>=1.0,<2",
    "sqlalchemy>=2.0,<3",
    "uvicorn[standard]>=0.30,<1",
]
